refactor InitState to LiveState
